# PRMS-FIXED/docs/PRMS_OpenAPI_v1_0.yaml
openapi: 3.1.0
info:
  title: PRMS — Patient Referral Management System API
  version: 1.0.0
  description: |
    Production API for the Kenya Patient Referral Management System (PRMS).
    
    **Base URL:** `https://api.prms.health.go.ke/api/v1`
    
    **Authentication:** Bearer JWT (RS256). Obtain tokens via `/auth/login`.
    
    **Compliance:** Kenya Data Protection Act (KDPA) 2019.
  contact:
    name: PRMS Engineering Team
    email: engineering@prms.health.go.ke
  license:
    name: Proprietary
servers:
  - url: https://api.prms.health.go.ke/api/v1
    description: Production
  - url: https://staging-api.prms.health.go.ke/api/v1
    description: Staging
  - url: http://localhost:3000/api/v1
    description: Local Development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 signed JWT. Expires in 15 minutes.

# ─────────────────────────────────────────────────────────────────────────────
# REUSABLE SCHEMAS
# ─────────────────────────────────────────────────────────────────────────────
  schemas:
...
    Meta:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          format: uuid
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
    MetaWithPagination:
      allOf:
        - $ref: '#/components/schemas/Meta'
        - type: object
          properties:
            pagination:
              $ref: '#/components/schemas/Pagination'
    ErrorDetail:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
            details:
              type: array
              items:
                $ref: '#/components/schemas/ErrorDetail'
        meta:
          $ref: '#/components/schemas/Meta'

    # ── Auth ─────────────────────────────────────────────────────────────────
    UserSummary:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        fullName:
          type: string
        role:
          type: string
          enum: [System Admin, Hospital Admin, Clinician, Receptionist]
        hospitalId:
          type: integer
          nullable: true
        hospitalName:
          type: string
          nullable: true
        isFirstLogin:
          type: boolean
    AuthTokens:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          description: Access token lifetime in seconds
          example: 900
        user:
          $ref: '#/components/schemas/UserSummary'

    # ── Hospital ──────────────────────────────────────────────────────────────
...
    PatientMasked:
      type: object
      description: Patient record with PII masked (Receptionist view)
      properties:
        id:
          type: integer
        fullName:
          type: string
          example: Jane W. M.
        nationalId:
          type: string
          example: XXXX6789
        gender:
          type: string
          enum: [Male, Female, Other]
        age:
          type: integer
        county:
          type: string
        phoneNumber:
          type: string
          example: 071XXXX678
    PatientFull:
      type: object
      description: Full patient record (Clinician view)
      properties:
        id:
          type: integer
        fullName:
          type: string
          example: Jane Wambui Mwangi
        nationalId:
          type: string
          example: "23456789"
        gender:
          type: string
          enum: [Male, Female, Other]
        dateOfBirth:
          type: string
          format: date
        age:
          type: integer
        county:
          type: string
        phoneNumber:
          type: string
          example: "+254712345678"
        createdAt:
          type: string
          format: date-time

    # ── Referral ──────────────────────────────────────────────────────────────
...
  responses:
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: AUTH_TOKEN_EXPIRED
              message: Access token has expired
            meta:
              timestamp: "2026-06-10T10:00:00.000Z"
              requestId: uuid
...
    ValidationError:
      description: Request body validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
paths:
...
  /auth/verify-2fa:
    post:
      tags: [Authentication]
      summary: Complete 2FA verification
      operationId: verify2FA
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [preAuthToken, otpCode]
              properties:
                preAuthToken:
                  type: string
                otpCode:
                  type: string
                  pattern: '^\d{6}$'
                  example: "584920"
      responses:
        '200':
          description: Verification successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/AuthTokens'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      accessToken:
                        type: string
                      expiresIn:
                        type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /auth/forgot-password:
    post:
      tags: [Authentication]
      summary: Request password reset email
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Always returns 200 to prevent email enumeration
  /auth/reset-password:
    post:
      tags: [Authentication]
      summary: Complete password reset
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [resetToken, newPassword, confirmPassword]
              properties:
                resetToken:
                  type: string
                newPassword:
                  type: string
                  format: password
                  minLength: 12
                confirmPassword:
                  type: string
                  format: password
      responses:
        '200':
          description: Password reset successfully
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /auth/change-password:
    patch:
      tags: [Authentication]
      summary: Change own password
      operationId: changePassword
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword, confirmPassword]
              properties:
                currentPassword:
                  type: string
                  format: password
                newPassword:
                  type: string
                  format: password
                  minLength: 12
                confirmPassword:
                  type: string
                  format: password
      responses:
        '200':
          description: Password changed
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/register-device:
    post:
      tags: [Authentication]
      summary: Register mobile device for push notifications
      operationId: registerDevice
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fcmToken, deviceId, platform]
              properties:
                fcmToken:
                  type: string
                deviceId:
                  type: string
                  format: uuid
                platform:
                  type: string
                  enum: [android, ios]
      responses:
        '200':
          description: Device registered

  # ═══════════════════════════════════════════════════════
  # HOSPITALS
  # ═══════════════════════════════════════════════════════
...
  /patients:
    post:
      tags: [Patients]
      summary: Register a new patient
      operationId: createPatient
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [idType, nationalId, fullName, gender, dateOfBirth, county]
              properties:
                idType:
                  type: string
                  enum: [National ID, Alien ID, Birth Certificate]
                nationalId:
                  type: string
                fullName:
                  type: string
                gender:
                  type: string
                  enum: [Male, Female, Other]
                dateOfBirth:
                  type: string
                  format: date
                county:
                  type: string
                phoneNumber:
                  type: string
      responses:
        '201':
          description: Patient registered
        '409':
          description: Patient with this ID already registered

    get:
      tags: [Patients]
      summary: Search patients
      operationId: searchPatients
      security:
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: National ID or partial name
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Patient list (role-masked)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      oneOf:
                        - $ref: '#/components/schemas/PatientFull'
                        - $ref: '#/components/schemas/PatientMasked'
                  meta:
                    $ref: '#/components/schemas/MetaWithPagination'
...